Improve SIP packet detection. Fixes #42
Some packets with payload with Call-Id headers were being considered
SIP packets, and that may no be true.

For example, homer packets or moving a sip pcap file to a NFS will
create packets with Call-Id in its payload. Improve the regexp for
matching METHOD and REQUEST CODE to ensure the payload is an actual
SIP message.