Remove ignore configuration directive.
Ignore directive is not well coded. In order to work properly it
requires message attributes to exists for checking their values.
This forces payload parsing of each SIP messages with extra overload
and memory usage.

Not sure about the impact of removing this feature, but we must think
other way to implement it.