Call Flow: Compressed mode using alias
Now compressed mode (pressing `s` in call flow) won't display ports
in column addresses. Instead, the alias of the address will be displayed
(if configured in `sngreprc` file)

Addresses configured with the same alias will be displayed as one unique
column. This may lead to confusion in some cases, but it also allows
a more clear representation of hosts with multiple addresses.