Disable RTP arrows in Call Flow by default
Change default value of cf.media setting to off.

Flows that have RTP streams will have the toggle streams keybinding
"highlighted" (in yellow).

This aims to improve the visibility of SIP messages when lots of arrows
are being displayed.

Of course, if cf.media is enabled in sngreprc resource file you can
obtain the previous behaviour.