Keystroke propagation to previous panel/s
The motivation of this feature is being able to run an action which is not
handled (or is only partially handled) by the current panel. After
handling the key on the current panel, this is destroyed by the ui_manager
and the key is passed to the previous panel. This can be done recursively until
the last panel is reached.

A clear use case is the ability to invoke ACTION_CLEAR_CALLS action from any of
'call_list', 'call_flow', or 'call_raw' with a single keystroke.