config: remove capture.rtp and sip.calls from settings panel
The main reason to remove this entries from configuration panel is
that are usually configured from command line. That may lead to an
unspected behaviour if you start with -rc command line flags and
save the configuration from settings panel: those settings will
be saved, not with their default value but with the changed from
the command line.