cf: rework how RTP arrow are displayed.
Now SIP message arrows store their source and destination columns.

For the RTP arrow, if the RTP packets addresses matches the source
address of the messages that contains the SDP that configured that
stream, that column will be used. If not, any column with that
displays the IP address will be used.