configure.ac: switch to pkg-config to find openssl
Using AC_CHECK_LIB() doesn't work properly for static linking, because
it doesn't return information on transitive dependencies of
libraries. For example, if library A internally uses library B, then
with dynamic linking, doing -lA is sufficient. However, with static
linking, one must do -lA -lB, and AC_CHECK_LIB() will only give the -lA
information. This for example causes a build failure when building
sngrep statically with openssl enabled:

checking for SSL_new in -lssl... no
configure: error:  You need to have libssl installed to compile sngrep

due to undefined symbols in the OpenSSL library.

The proper solution for this is to discover the library using
pkg-config, because pkg-config properly returns the list of necessary
libraries, as it understands the concept of "list of libraries needed
when dynamic linking" and "list of libraries needed for static linking".

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>