cl: fix buffer overflow in display filter #225
Display filter content was being moved into a temporal
buffer with fixed length of 256. The filter field is
determined by the actual screen width, so we must ensure
the buffer has enough space to contain it.

ncurses always returns the same length output for the
filter value, filling the buffer with whitespaces until the
field lenght has been reached.