replace #elifdef with #elif defined()
The #elifdef preprocessor directive will only become available with the
upcoming C2x standard.

I do not think that using it already is a good idea, since support for
it is very sparse and no long-term-stable distro has support for it.
Also replacing it with "#elif defined()" doesn't really complicate the code.