Do proper shutdown upon SIGTERM, SIGINT or SIGQUIT
Before this patch the process was just terminated by the default signal
handler. This could lead for example to incompletely written pcap files.
So proper shutdown like in this patch is always a good idea. But it becomes
mandadory when implementing more complex file I/O with buffering, for
example with gzip streaming.

To not conflict with the threading used in sngrep, the signal handler
just sets a atomic flag. This flag is then checked in one of the two
main loops (with/without curses UI).