sip: Manage active call list properly.
As the `sip_call_is_active` function is checking the active call list
for the given call, it is preventing newly created calls from ever
being added to the active list.

Additionally, `vector_remove` returns silently when the given item is
not present, so there is no reason to check if the call is in the
active list before trying to remove it.